fixup! Enh: Add Application interface and make sure the `onInit` event is fired in the web application too (#6505)
